../../../../ext/compressors/bzip2_compress/bzip2_compress.c:161: warning: implicit conversion shortens 64-bit value into a 32-bit value
--HG--
extra : rebase_source : 67f5d6fe20895f7f2891df98c1062b84a2db85da

+++ ext/compressors/bzip2_compress/bzip2_compress.c
@@ -158,9 +158,9 @@ bzip2_compress(WT_COMPRESSOR *compressor, WT_SESSION *session,
 		return (bzip2_error(session, "BZ2_bzCompressInit", ret));
 
 	bz.next_in = (char *)src;
-	bz.avail_in = src_len;
+	bz.avail_in = (uint32_t)src_len;
 	bz.next_out = (char *)dst;
-	bz.avail_out = dst_len;
+	bz.avail_out = (uint32_t)dst_len;
 	if ((ret = BZ2_bzCompress(&bz, BZ_FINISH)) == BZ_STREAM_END) {
 		*compression_failed = 0;
 		*result_lenp = dst_len - bz.avail_out;
@@ -193,9 +193,9 @@ bzip2_decompress(WT_COMPRESSOR *compressor, WT_SESSION *session,
 		return (bzip2_error(session, "BZ2_bzDecompressInit", ret));
 
 	bz.next_in = (char *)src;
-	bz.avail_in = src_len;
+	bz.avail_in = (uint32_t)src_len;
 	bz.next_out = (char *)dst;
-	bz.avail_out = dst_len;
+	bz.avail_out = (uint32_t)dst_len;
 	if ((ret = BZ2_bzDecompress(&bz)) == BZ_STREAM_END) {
 		*result_lenp = dst_len - bz.avail_out;
 		ret = 0;
